Apply reviewer comment
