+++ common/base/src/lib.rs
@@ -24,6 +24,7 @@ mod stoppable_test;
 mod profiling;
 mod progress;
 mod runtime;
+mod shutdown_signal;
 mod stop_handle;
 mod stoppable;
 mod uniq_id;
@@ -36,6 +37,9 @@ pub use runtime::BlockingWait;
 pub use runtime::Dropper;
 pub use runtime::Runtime;
 pub use runtime::TrySpawn;
+pub use shutdown_signal::signal_stream;
+pub use shutdown_signal::SignalStream;
+pub use shutdown_signal::SignalType;
 pub use stop_handle::StopHandle;
 pub use stoppable::Stoppable;
 pub use tokio;
